agent/requirements.txt
# 核心依赖
langchain-docling>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
sqlite3  # Python内置，无需安装

# AI和机器学习
langchain-google-vertexai>=1.0.0
langchain-core>=0.3.0
langchain>=0.3.0
transformers>=4.30.0
torch>=2.0.0

# 数据处理
pandas>=1.5.0
numpy>=1.24.0
json5>=0.9.0

# 网络和下载
requests>=2.28.0
urllib3>=1.26.0

# 日志和工具
logging  # Python内置
typing-extensions>=4.5.0
python-dotenv>=1.0.0

# 可选：Jupyter支持
jupyter>=1.0.0
ipykernel>=6.0.0

# 可选：进度条
tqdm>=4.65.0 